product_service/requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.5.0
pillow==10.2.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==3.12.3
python-dotenv==1.0.1
pytz==2023.4
sqlparse==0.2.4
typing_extensions==4.11.0
